+++ application/backend/controllers/home.js
@@ -13,14 +13,15 @@ HomeController.prototype.routes = function event(request, reply) {
             '/drug/enforcement.json',
             '/device/event.json',
             '/device/enforcement.json',
-            '/food/enforcement.json',
-            '/goodrx/fair-price.json',
-            '/goodrx/low-price.json',
-            '/goodrx/compare-price.json',
-            '/goodrx/drug-info.json',
-            '/goodrx/drug-search.json'
+            '/food/enforcement.json'
         ]
     })
 };
+
+//'/goodrx/fair-price.json',
+//'/goodrx/low-price.json',
+//'/goodrx/compare-price.json',
+//'/goodrx/drug-info.json',
+//'/goodrx/drug-search.json'
 
 module.exports = new HomeController();
