+++ application/backend/routes.js
@@ -6,7 +6,7 @@ var foodController = require('./controllers/food');
 var symptomController = require('./controllers/symptom');
 var homeController = require('./controllers/home');
 var productController = require('./controllers/product');
-var goodRxController = require('./controllers/goodrx');
+//var goodRxController = require('./controllers/goodrx');
 
 exports.endpoints = [
     {
@@ -72,41 +72,41 @@ exports.endpoints = [
             handler: productController.search
             }
     },
-    {
-        method: 'GET',
-        path: '/goodrx/fair-price.json',
-        config: {
-            handler: goodRxController.fairprice
-        }
-    },
-    {
-        method: 'GET',
-        path: '/goodrx/low-price.json',
-        config: {
-            handler: goodRxController.lowprice
-        }
-    },
-    {
-        method: 'GET',
-        path: '/goodrx/compare-price.json',
-        config: {
-            handler: goodRxController.compareprice
-        }
-    },
-    {
-        method: 'GET',
-        path: '/goodrx/drug-info.json',
-        config: {
-            handler: goodRxController.druginfo
-        }
-    },
-    {
-        method: 'GET',
-        path: '/goodrx/drug-search.json',
-        config: {
-            handler: goodRxController.drugsearch
-        }
-    },
+    //{
+    //    method: 'GET',
+    //    path: '/goodrx/fair-price.json',
+    //    config: {
+    //        handler: goodRxController.fairprice
+    //    }
+    //},
+    //{
+    //    method: 'GET',
+    //    path: '/goodrx/low-price.json',
+    //    config: {
+    //        handler: goodRxController.lowprice
+    //    }
+    //},
+    //{
+    //    method: 'GET',
+    //    path: '/goodrx/compare-price.json',
+    //    config: {
+    //        handler: goodRxController.compareprice
+    //    }
+    //},
+    //{
+    //    method: 'GET',
+    //    path: '/goodrx/drug-info.json',
+    //    config: {
+    //        handler: goodRxController.druginfo
+    //    }
+    //},
+    //{
+    //    method: 'GET',
+    //    path: '/goodrx/drug-search.json',
+    //    config: {
+    //        handler: goodRxController.drugsearch
+    //    }
+    //},
     {
         method: 'GET',
         path: '/',
